rename: 파일 관련 API DTO 및 디렉토리 변경
