+++ packy-api/src/main/java/com/dilly/admin/api/FileController.java
@@ -1,7 +1,7 @@
-package com.dilly.file;
+package com.dilly.admin.api;
 
+import com.dilly.admin.dto.response.UrlResponse;
 import com.dilly.application.FileService;
-import com.dilly.dto.request.FileRequest;
 import com.dilly.exception.ErrorCode;
 import com.dilly.global.response.DataResponseDto;
 import com.dilly.global.swagger.ApiErrorCodeExample;
@@ -28,12 +28,14 @@ public class FileController {
 	)
 	@GetMapping("/presigned-url/{fileName}")
 	@ApiErrorCodeExample(ErrorCode.FILE_SERVER_ERROR)
-	public DataResponseDto<FileRequest> getPresignedUrl(
+	public DataResponseDto<UrlResponse> getPresignedUrl(
 		@PathVariable(name = "fileName") @Schema(description = "확장자명을 포함해주세요")
 		String fileName) {
 		String activeProfile = System.getProperty("spring.profiles.active");
 
-		return DataResponseDto.from(
+		UrlResponse urlResponse = UrlResponse.from(
 			fileService.getPresignedUrl("images/" + activeProfile, fileName));
+
+		return DataResponseDto.from(urlResponse);
 	}
 }
